Move test functions in `test_study` files to `test_optimize.py`
